requirements-all.txt
# =============================================================================
# Selgis ML - Full Requirements
# =============================================================================
# Usage: pip install -r requirements-all.txt
# =============================================================================

# Base requirements
-r requirements.txt

# -----------------------------------------------------------------------------
# TRANSFORMERS AND LLM
# -----------------------------------------------------------------------------
transformers>=4.30.0
datasets>=2.14.0
accelerate>=0.21.0
tokenizers>=0.13.0

# -----------------------------------------------------------------------------
# LORA / PEFT / QUANTIZATION
# -----------------------------------------------------------------------------
peft>=0.5.0
bitsandbytes>=0.41.0
scipy>=1.10.0

# -----------------------------------------------------------------------------
# IMAGES AND MULTIMODAL
# -----------------------------------------------------------------------------
pillow>=9.0.0
torchvision>=0.15.0

# -----------------------------------------------------------------------------
# LOGGING
# -----------------------------------------------------------------------------
wandb>=0.15.0

# -----------------------------------------------------------------------------
# UTILS
# -----------------------------------------------------------------------------
pyyaml>=6.0
